Minor update
Fixed previously unknown issues with the player attack zone physics bodies (they existed too long, had gravity and were effected by kinetic force) and edited the ball vs player (player death logics.) There's a velocity check in there now but it's currently a bit finicky; needs more work. I also removed the "stupidBall.png" that I accidently put in the first commit. It was a square and was going to be a temp asset but I decided to take the 1 minutes necessary to make a circle.